requirementsWithoutFakeRpi.txt
appdirs==1.4.4
backcall==0.2.0
betterproto==1.2.5
black==19.10b0
CacheControl==0.12.6
cachy==0.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cleo==0.8.1
click==7.1.2
clikit==0.6.2
crashtest==0.3.1
cryptography==3.4.6
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
grpclib==0.3.2
h2==4.0.0
hpack==4.0.0
html5lib==1.1
hyperframe==6.0.0
idna==2.10
iniconfig==1.1.1
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
jeepney==0.6.0
Jinja2==2.11.3
jsonpickle==2.0.0
keyring==21.8.0
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==1.0.2
multidict==5.1.0
paho-mqtt==1.5.1
parso==0.8.1
pastel==0.2.1
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
pkginfo==1.7.0
poetry==1.1.5
poetry-core==1.0.2
prompt-toolkit==3.0.18
protobuf==3.15.6
psutil==5.8.0
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
pylev==1.3.0
regex==2021.3.17
requests==2.25.1
requests-toolbelt==0.9.1
SecretStorage==3.3.1
shellingham==1.4.0
six==1.15.0
stringcase==1.2.0
toml==0.10.2
tomlkit==0.7.0
traitlets==5.0.5
typed-ast==1.4.2
typing==3.7.4.3
urllib3==1.26.4
virtualenv==20.4.3
webencodings==0.5.1
vcgencmd==0.1.1
